# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Delivery",
        "description": "API del Sistema de Gestión de Delivery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/comprobantes": {
            "get": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Listar todos los comprobantes",
                "operationId": "f7409aef67dd9609bbc649481c7a6258",
                "responses": {
                    "200": {
                        "description": "Comprobantes obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay comprobantes disponibles."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            },
            "post": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Crear un nuevo comprobante",
                "operationId": "bb2cbfdb9cc216637300c8d51bcf862f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pago_id",
                                    "tipo_comprobante",
                                    "numero_comprobante",
                                    "titular"
                                ],
                                "properties": {
                                    "pago_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_comprobante": {
                                        "type": "string",
                                        "example": "Factura A",
                                        "maxLength": 50
                                    },
                                    "numero_comprobante": {
                                        "type": "string",
                                        "example": "0001-00000123",
                                        "maxLength": 100
                                    },
                                    "titular": {
                                        "type": "string",
                                        "example": "Juan Pérez",
                                        "maxLength": 150
                                    },
                                    "url_pdf": {
                                        "type": "string",
                                        "example": "https://ejemplo.com/comprobante.pdf",
                                        "nullable": true,
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comprobante creado correctamente."
                    },
                    "422": {
                        "description": "Datos inválidos o error de validación."
                    },
                    "500": {
                        "description": "Error interno al crear el comprobante."
                    }
                }
            }
        },
        "/api/comprobantes-desactivados": {
            "get": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Listar todos los comprobantes desactivados",
                "operationId": "76e5cdf2f3b5592c9eb821beb9da323b",
                "responses": {
                    "200": {
                        "description": "Comprobantes desactivados obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay comprobantes desactivados para mostrar."
                    },
                    "505": {
                        "description": "Error interno del servidor al obtener los comprobantes desactivados."
                    }
                }
            }
        },
        "/api/comprobantes/{id}": {
            "get": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Obtener un comprobante específico",
                "operationId": "50166a550235027955e71412c004b2f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comprobante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprobante obtenido correctamente."
                    },
                    "404": {
                        "description": "Comprobante no encontrado."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Eliminar un comprobante por ID",
                "operationId": "f7de9fc984ccd55e1283825218251190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comprobante a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprobante eliminado correctamente."
                    },
                    "404": {
                        "description": "Comprobante no encontrado."
                    },
                    "500": {
                        "description": "Error interno al eliminar el comprobante."
                    }
                }
            }
        },
        "/api/comprobantes/{id}/restore": {
            "put": {
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Restaurar un comprobante eliminado",
                "operationId": "36661adb089a0f81ed871a79c94a614e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del comprobante a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprobante restaurado correctamente."
                    },
                    "404": {
                        "description": "Comprobante no encontrado entre los eliminados."
                    },
                    "500": {
                        "description": "Error interno al restaurar el comprobante."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}